+++ async.stub.php
@@ -33,7 +33,7 @@ function suspend(): void {}
 /**
  * Execute the provided closure in non-cancellable mode.
  */
-function protect(\Closure $closure): void {}
+function protect(\Closure $closure): mixed {}
 
 function await(Awaitable $awaitable, ?Awaitable $cancellation = null): mixed {}
 
@@ -41,13 +41,13 @@ function awaitAny(iterable $triggers, ?Awaitable $cancellation = null): mixed {}
 
 function awaitFirstSuccess(iterable $triggers, ?Awaitable $cancellation = null): mixed {}
 
-function awaitAll(iterable $triggers, ?Awaitable $cancellation = null, bool $fillNull = false): array {}
+function awaitAll(iterable $triggers, ?Awaitable $cancellation = null, bool $preserveKeyOrder = true): array {}
 
-function awaitAllWithErrors(iterable $triggers, ?Awaitable $cancellation = null, bool $fillNull = false): array {}
+function awaitAllWithErrors(iterable $triggers, ?Awaitable $cancellation = null, bool $preserveKeyOrder = true, bool $fillNull = false): array {}
 
-function awaitAnyOf(int $count, iterable $triggers, ?Awaitable $cancellation = null): array {}
+function awaitAnyOf(int $count, iterable $triggers, ?Awaitable $cancellation = null, bool $preserveKeyOrder = true): array {}
 
-function awaitAnyOfWithErrors(int $count, iterable $triggers, ?Awaitable $cancellation = null): array {}
+function awaitAnyOfWithErrors(int $count, iterable $triggers, ?Awaitable $cancellation = null, bool $preserveKeyOrder= true, bool $fillNull = false): array {}
 
 function delay(int $ms): void {}